pyproject.toml
[tool.poetry]
name = "chemist"
version = "1.4.1"
description = "A simple, flexible and testable active-record powered by SQLAlchemy."
authors = ["Gabriel Falcão <gabriel@nacaolivre.org>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.6"
ipdb = "^0.12.0"
plant = "^0.1.3"
pynacl = "^1.3"
python-dateutil = "^2.8"
pyyaml = "^5.1"
sqlalchemy = "^1.3"

[tool.poetry.dev-dependencies]
pathlib2 = "^2.3"
nose = "^1.3"
pinocchio = "^0.4.2"
sphinx = "^2.0"
sphinx-rtd-theme = "^0.4.3"
sure = "^1.4"
twine = "^1.13"
rednose = "^1.3"
mock = "^3.0"
coverage = "^4.5"
mysqlclient = "^1.4"
ipdb = "^0.12.0"
psycopg2-binary = "^2.8"
"testing.postgresql" = "^1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
